Fix changes
pick 7917b57 Implementation of two new models: Logistic Regression and Naive Bayes instead of Random Forest Randomized Search CV. Cross Validation of the selected model (KNN) and interpretation of the results based on the error matrix. Selection of hyperparameters of the final model. Model training on new data (5 variables correlated with the target) and trying out the prediction. Adding descriptions and conclusions regarding the entire project.